Move authentication logic to a spearate file
